Reorganized folders for new project box series
Major changes to the membership and subscription model and the addition of 4 primary project box products necessitated a change in organization of these folders to avoid confution.  The Young Engineer Series is still happening, but in a different form.  All of the old project box files are still here in a subfolder.